Initial project

+++ src/Authentication/Authentication.js
@@ -3,9 +3,7 @@ import "./Authentication.css";
 import {useNavigate} from "react-router";
 import {useState} from 'react';
 import authCorner1 from "../Pictures/authCorner1.png";
-import authCorner2 from "../Pictures/authCorner2.png";
 import authCorner3 from "../Pictures/authCorner3.png";
-import authCorner4 from "../Pictures/authCorner4.png";
 import authWhiteFlower from "../Pictures/authWhiteFlower.png";
 import authRose from "../Pictures/authRose.png";
 
@@ -13,12 +11,13 @@ const Authentication = () => {
     const navigate = useNavigate();
     const [loginEmail, setLoginEmail] = useState("");
     const [loginPassword, setLoginPassword] = useState("");
-    const [isLoginError, setIsLoginError] = useState(false);
+    const [loginError, setLoginError] = useState(null);
 
     const [registerEmail, setRegisterEmail] = useState("");
     const [registerPassword, setRegisterPassword] = useState("");
     const [registerName, setRegisterName] = useState("");
-    const [isRegisterError, setIsRegisterError] = useState(false);
+    const [registerConfirmPassword, setRegisterConfirmPassword] = useState("");
+    const [registerError, setRegisterError] = useState(null);
 
     const [isSignUp, setIsSignUp] = useState(false);
 
@@ -31,51 +30,63 @@ const Authentication = () => {
     }
 
     const handleLoginClick = async (event) => {
-        event.preventDefault(); // prevent form submission
+        if (!loginEmail || !loginPassword) {
+            setLoginError("Email and password required")
+        } else {
+            event.preventDefault(); // prevent form submission
 
-        try {
-            const response = await fetch('http://localhost:3001/login', {
-                method: 'POST',
-                headers: {
-                    'Content-Type': 'application/json',
-                },
-                body: JSON.stringify({email: loginEmail, password: loginPassword}),
-            });
+            try {
+                const response = await fetch('http://localhost:3001/login', {
+                    method: 'POST',
+                    headers: {
+                        'Content-Type': 'application/json',
+                    },
+                    body: JSON.stringify({email: loginEmail, password: loginPassword}),
+                });
 
-            if (response.status === 200) {
-                sessionStorage.setItem("loggedInUser", loginEmail);
-                navigate("/main");
-            } else if (response.status === 401) {
-                setIsLoginError(true)
+                if (response.status === 200) {
+                    sessionStorage.setItem("loggedInUser", loginEmail);
+                    navigate("/main");
+                } else if (response.status === 401) {
+                    setLoginError("Wrong username or password")
+                }
+            } catch (error) {
+                console.error('Error while fetching data', error);
             }
-
-        } catch (error) {
-            console.error('Error while fetching data', error);
         }
     }
 
     const handleRegisterClick = async (event) => {
-        event.preventDefault(); // prevent form submission
+        if (!registerEmail || !registerPassword || !registerName || !registerConfirmPassword) {
+            setRegisterError("All values are required!")
+        } else if (registerPassword !== registerConfirmPassword) {
+            setRegisterError("Passwords not matching!")
+        } else if (registerPassword.length < 6) {
+            setRegisterError("Password must be at least 6 characters!");
+        } else {
+            event.preventDefault(); // prevent form submission
 
-        try {
-            const response = await fetch('http://localhost:3001/register', {
-                method: 'POST',
-                headers: {
-                    'Content-Type': 'application/json',
-                },
-                body: JSON.stringify({email: registerEmail, password: registerPassword, name: registerName}),
-            });
+            try {
+                const response = await fetch('http://localhost:3001/register', {
+                    method: 'POST',
+                    headers: {
+                        'Content-Type': 'application/json',
+                    },
+                    body: JSON.stringify({email: registerEmail, password: registerPassword, name: registerName}),
+                });
 
-            if (response.status === 200) {
-                sessionStorage.setItem("loggedInUser", registerEmail);
-                navigate("/main");
-            } else if (response.status === 409) {
-                setIsRegisterError(true)
+                if (response.status === 200) {
+                    sessionStorage.setItem("loggedInUser", registerEmail);
+                    navigate("/main");
+                } else if (response.status === 409) {
+                    setRegisterError("User already registered!")
+                }
+
+            } catch (error) {
+                console.error('Error while fetching data', error);
             }
-
-        } catch (error) {
-            console.error('Error while fetching data', error);
         }
+
     }
 
     const pinkBoxStyle = {
@@ -99,39 +110,45 @@ const Authentication = () => {
                 <div className="pinkbox" style={pinkBoxStyle}>
                     <div className={signUpClassName}>
                         <h1 className="authH1">register</h1>
-                        <form className="authForm" autoComplete="off">
+                        <div className="authForm">
                             <input className="authInput" type="text" placeholder="username"
                                    onChange={(e) => {
                                        setRegisterName(e.target.value)
-                                       setIsRegisterError(false)
+                                       setRegisterError(false)
                                    }
                                    }/>
                             <input className="authInput" type="email" placeholder="email"
                                    onChange={(e) => {
                                        setRegisterEmail(e.target.value)
-                                       setIsRegisterError(false)
+                                       setRegisterError(false)
                                    }}/>
                             <input className="authInput" type="password" placeholder="password"
                                    onChange={(e) => {
                                        setRegisterPassword(e.target.value)
-                                       setIsRegisterError(false)
+                                       setRegisterError(false)
                                    }}/>
-                            <input className="authInput" type="password" placeholder="confirm password"/>
-                            {isRegisterError ? <div>User already exists!</div> : null}
+                            <input
+                                className="authInput"
+                                onChange={(e) => {
+                                    setRegisterConfirmPassword(e.target.value)
+                                    setRegisterError(false)
+                                }}
+                                type="password" placeholder="confirm password"/>
+                            {registerError ? registerError : null}
                             <button onClick={handleRegisterClick} className="authButton authButton2 authSubmit">create
                                 account
                             </button>
-                        </form>
+                        </div>
                     </div>
                     <div className={signInClassName}>
                         <h1 className="authH1">sign in</h1>
-                        <form className="more-padding authForm" autoComplete="off">
+                        <div className="more-padding authForm">
                             <input
                                 className="authInput"
                                 type="text"
                                 onChange={(e) => {
                                     setLoginEmail(e.target.value)
-                                    setIsLoginError(false)
+                                    setLoginError(null)
                                 }}
                                 placeholder="email"/>
                             <input className="authInput"
@@ -139,30 +156,26 @@ const Authentication = () => {
                                    placeholder="password"
                                    onChange={(e) => {
                                        setLoginPassword(e.target.value)
-                                       setIsLoginError(false)
+                                       setLoginError(null)
                                    }}/>
-                            <div className="authCheckbox">
-                                <input className="authInput" type="checkbox" id="remember"/>
-                                <label className="authLabel" htmlFor="remember">remember me</label>
-                            </div>
 
-                            {isLoginError ? <div>Wrong email or password</div> : null}
+                            {loginError ? loginError : null}
                             <button onClick={handleLoginClick} className="authButton authButton2 authSubmit">log in
                             </button>
-                        </form>
+                        </div>
                     </div>
                 </div>
                 <div className="leftbox">
-                    <h2 className="authTitle"><span className="authSpan">BLOOM</span>&<br/>BOUQUET</h2>
-                    <p className="authDesc authP">pick your perfect <span className="authSpan">bouquet</span></p>
+                    <h2 className="authTitle"><span className="authSpan">Iconic Dream</span><br/>Occasions</h2>
+                    <p className="authDesc authP">make your perfect <span className="authSpan">wedding</span></p>
                     <img className="smaller" src={authRose}
                          alt="1357d638624297b" border="0"/>
                     <p className="account authP">have an account?</p>
                     <button className="authButton authButton2" id="signIn" onClick={handleSignIn}>login</button>
                 </div>
                 <div className="rightbox">
-                    <h2 className="authTitle"><span className="authSpan">BLOOM</span>&<br/>BOUQUET</h2>
-                    <p className="authDesc authP"> pick your perfect <span className="authSpan">bouquet</span></p>
+                    <h2 className="authTitle"><span className="authSpan">Iconic Dream</span><br/>Occasions</h2>
+                    <p className="authDesc authP"> make your perfect <span className="authSpan">wedding</span></p>
                     <img alt="" className="flower" src={authWhiteFlower}/>
                     <p className="account authP">don't have an account?</p>
                     <button className="authButton authButton2" id="signUp" onClick={handleSignUp}>sign up</button>
